DemolitionCharge and BasementType moved to own files
