+++ megamek/src/megamek/common/Hex.java
@@ -18,8 +18,8 @@ import java.io.Serializable;
 import java.util.HashMap;
 import java.util.StringTokenizer;
 
-import megamek.common.Building.BasementType;
 import megamek.common.annotations.Nullable;
+import megamek.common.building.BasementType;
 
 /**
  * Hex represents a single hex on the board.
