+++ megamek/src/megamek/server/Server.java
@@ -82,8 +82,6 @@ import megamek.common.Board;
 import megamek.common.BoardDimensions;
 import megamek.common.BombType;
 import megamek.common.Building;
-import megamek.common.Building.BasementType;
-import megamek.common.Building.DemolitionCharge;
 import megamek.common.BuildingTarget;
 import megamek.common.CalledShot;
 import megamek.common.CommonConstants;
@@ -211,6 +209,8 @@ import megamek.common.actions.UnjamAction;
 import megamek.common.actions.UnjamTurretAction;
 import megamek.common.actions.UnloadStrandedAction;
 import megamek.common.actions.WeaponAttackAction;
+import megamek.common.building.BasementType;
+import megamek.common.building.DemolitionCharge;
 import megamek.common.containers.PlayerIDandList;
 import megamek.common.event.GameListener;
 import megamek.common.event.GameVictoryEvent;
